[WIP] [NOTES] cca_in_ad -> replace kubernetes/docker launchers with container launcher
The kubernetes launcher _only exists_ to handle `container_collect_all`;
with that handled by AD, there's no need for the launcher.  The docker
launcher's functionality is superseded by the container launcher.

The container launcher covers everything that the k8s and docker
launchers do, without some of their bugs.  In particular, it

 - handles config from files (conf.d) for containers under kubernetes
 - handles changes to pod annotations
 - correctly logs from file or socket based on `logs_config.{k8s,docker}_container_use_file`
 - captures logs under containerd when `k8s_container_use_file` is false

Outside of the situations where they have obvious bugs, this launcher
attempts to preserve behavior of the kubernetes and docker launchers in
the details, such as the "Source" and "Service" properties.  It does not
preserve those launchers' behavior with respect to `agent status`
output, as that behavior was not very intuitive.

In this commit, the container launcher includes its own, private notion
of a "tailer" which wraps existing logs-agent functionality: for
sockets, it wraps the socket tailer; for files, it wraps file _sources_.
Both of these wrappers are ugly and awful but allow reuse of the
existing functionality without modifying it.

The entire container launcher is hidden behind the 'docker' build tag,
which really means "containers" and not specifically docker.